Database connection.
Completing the connection to the database. Added a simple PHP script that selects all the different countries and outs them into a select tag. This is to test that the connection to the database is working.

Might make the HTML easier to read by creating different PHP files that does different things.